problems/src/test/java/_1001_1050/_1040_Moving_Stones_Until_Consecutive_II/SolutionITest.java
package _1001_1050._1040_Moving_Stones_Until_Consecutive_II;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

/**
 * @author jacka
 * @version 1.0 on 10/24/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final int[] input = {7, 4, 9}, expected = {1, 2}, actual = solution.numMovesStonesII(input);
    assertArrayEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final int[] input = {6, 5, 4, 3, 10}, expected = {2, 3},
        actual = solution.numMovesStonesII(input);
    assertArrayEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase3(Solution solution) {
    final int[] input = {100, 101, 104, 102, 103}, expected = {0, 0}, actual = solution.numMovesStonesII(input);
    assertArrayEquals(expected, actual);
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}